% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_csv.R
\name{exp_csv}
\alias{exp_csv}
\title{Exports a data table to a CSV with various options}
\usage{
exp_csv(
  .tbl,
  .file_pref,
  .download_dir = "~",
  .tidy_nms = 0,
  .glimpse = FALSE,
  .print = FALSE,
  .open = TRUE,
  .suffix_pref = FALSE,
  .incl_date = TRUE
)
}
\arguments{
\item{.tbl}{Tibble or data table to export}

\item{.file_pref}{The name of the destination file}

\item{.download_dir}{The destination of the exported file}

\item{.tidy_nms}{Enter 1 to tidy the column names (e.g. replace "_" with a space and set to title case)}

\item{.glimpse}{Enter TRUE to print a glimpse of the tibble while exporting to a CSV.}

\item{.print}{Enter TRUE to print the tibble while exporting to a CSV}

\item{.open}{Enter TRUE to open the file in the default application}

\item{.suffix_pref}{Enter TRUE to append the .file_pref to the file name after the date suffix}

\item{.incl_date}{Enter TRUE to append a date/time suffix to the file name}
}
\value{
Prints a summary of the exported file
}
\description{
Exports a data table to a CSV with various options
}
\examples{
exp_csv(iris, "iris_exp", "~", .open = FALSE, .incl_date = FALSE)
}
