% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftoy.R
\name{ftoy}
\alias{ftoy}
\title{Función de juguete}
\usage{
ftoy(n = 100)
}
\arguments{
\item{n}{tamaño muestral para X y Y, por defecto, igual a 100.}
}
\value{
Relación de X vs Y y el histograma de la variable aleatoria X.
}
\description{
La función genera dos variables aleatorias X y Y, de distribución
normal con desviación estándar igual a 1 y medias 5 y 10,
respectivamente. La función utiliza estos datos para graficar
el diagrama de dispersión de X vs Y y el histograma de la variable
aleatoria X.
}
\details{
Detalle importante de ftoy.
}
