% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{gen_tbl}
\alias{gen_tbl}
\title{Generate a tibble}
\usage{
gen_tbl(n, types)
}
\arguments{
\item{n}{The number of rows in the tibble.}

\item{types}{The types for the columns in the data. The representation
follows the convention specified in the \code{readr}-package and the types
available are c = character, i = integer, d = double, l = logical, D = date.}
}
\value{
A tibble with \code{n} rows and \code{nchar(types)} columns.
}
\description{
Generates a tibble of \code{n} rows with user-specified column types.
}
\examples{
library(genr)
# A tibble with 42 rows and one character, one date and one integer column
gen_tbl(42, "cDi")

}
