% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{gen_tbl_cust}
\alias{gen_tbl_cust}
\title{Generate a tibble}
\usage{
gen_tbl_cust(n, gens)
}
\arguments{
\item{n}{The number of rows in the tibble.}

\item{gens}{A list of functions that take a single argument \code{n} and
which are used to generate the columns.}
}
\value{
A tibble with \code{n} rows and \code{length(gens)} columns.
}
\description{
Generates a tibble of \code{n} rows with user-specified columns.
}
\examples{
\donttest{
# A tibble with 7 rows and one column with logical values
gen_tbl_cust(7, list(function(n) sample(c(FALSE, TRUE), n, replace = TRUE)))
}

}
