% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_dfs.R
\name{retrieve_dfs}
\alias{retrieve_dfs}
\title{Retrieve objects stored in DFS files}
\usage{
retrieve_dfs(.objs, .dir)
}
\arguments{
\item{.objs}{Character vector of RDS files to retrieve}

\item{.dir}{Directory where RDS files are located}
}
\value{
Imports objects in RDS files. Prints a list of successful imports and unsuccessful imports
}
\description{
Retrieve objects stored in DFS files
}
