% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sel_file.R
\name{sel_file}
\alias{sel_file}
\title{Select a file, usually to import, and post a message with the file's date stamp}
\usage{
sel_file(.caption = "Select file to open", .path = "~", .filter = "*.xlsx")
}
\arguments{
\item{.caption}{Enter a prompt for the dialogue box displayed when selecting the file}

\item{.path}{Enter the default path to the file}

\item{.filter}{Enter the default file name pattern to streamline the file's selection}
}
\value{
Returns a string value with the selected file's path and name. Also prints the file's full path and a date stamp
}
\description{
Select a file, usually to import, and post a message with the file's date stamp
}
