% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dfs.R
\name{write_dfs}
\alias{write_dfs}
\title{Exports the specified objects to RDS files}
\usage{
write_dfs(.objs, .dir)
}
\arguments{
\item{.objs}{Character vector of the objects to export}

\item{.dir}{Directory to export to}
}
\value{
Objects are exported to DFS files and confimration printed
}
\description{
Exports the specified objects to RDS files
}
\examples{
write_dfs(c("mt_cars"), .dir = "~")
}
